+++ packages/components/src/DataViewer/Core/Tree/Tree.test.js
@@ -1,5 +1,25 @@
-import { shallow } from 'enzyme';
+/* eslint-disable react/prop-types */
+/* eslint-disable react/display-name */
+import { render, screen } from '@testing-library/react';
 import Tree, { isRoot } from './Tree.component';
+
+jest.mock('../TreeNode', () => props => (
+	<div className={props.className} data-testid="TreeNode" data-props={JSON.stringify(props)} />
+));
+jest.mock('../TreeNodeList', () => props => (
+	<ul
+		className={props.treeClassName}
+		data-testid="TreeNodeList"
+		data-props={JSON.stringify(props)}
+		data-level={props.level}
+	>
+		{props.value.map((item, index) => (
+			<li key={index} data-testid="TreeNodeListItem" className={props.nodeClassName}>
+				{item}
+			</li>
+		))}
+	</ul>
+));
 
 describe('isRoot', () => {
 	it('should return true, its root level', () => {
@@ -11,36 +31,36 @@ describe('isRoot', () => {
 });
 
 describe('Tree', () => {
-	let props;
-	beforeEach(() => {
-		props = {
-			getJSONPath: jest.fn(),
-			branch: jest.fn(),
-			leaf: jest.fn(),
-			getValueType: jest.fn(),
-			jsonpath: '$',
-		};
-	});
 	it('should return a TreeNodeList', () => {
-		const wrapper = shallow(<Tree {...props} value={[]} level={0} noRoot />);
-		expect(wrapper.getElement()).toMatchSnapshot();
+		render(<Tree value={[]} level={0} noRoot />);
+		expect(screen.getByRole('list')).toBeVisible();
+		expect(JSON.parse(screen.getByRole('list').dataset.props)).toMatchObject({
+			level: 0,
+			value: [],
+			treeClassName: 'theme-tc-tree tc-tree',
+		});
 	});
-	it('should return a TreeNode', () => {
-		const wrapper = shallow(<Tree {...props} value={{}} dataKey="myDataKey" level={0} />);
-		expect(wrapper.getElement()).toMatchSnapshot();
+	it('should return a TreeNode with dataKey', () => {
+		render(<Tree value={{}} dataKey="myDataKey" level={0} />);
+		expect(screen.getByTestId('TreeNode')).toBeVisible();
+		expect(JSON.parse(screen.getByTestId('TreeNode').dataset.props)).toMatchObject({
+			dataKey: 'myDataKey',
+			level: 0,
+		});
 	});
-	it('should return a TreeNodeList with custom className', () => {
-		const wrapper = shallow(
-			<Tree {...props} value={[]} className="myCustomClass" level={0} noRoot />,
-		);
-		expect(wrapper.getElement()).toMatchSnapshot();
+	it('should return a list with custom className', () => {
+		render(<Tree value={[]} className="myCustomClass" level={0} noRoot />);
+		expect(screen.getByRole('list')).toHaveClass('myCustomClass');
 	});
-	it('should return a TreeNodeList with a level > 0', () => {
-		const wrapper = shallow(<Tree {...props} value={[]} level={1} />);
-		expect(wrapper.getElement()).toMatchSnapshot();
+	it('should return a list with a level > 0', () => {
+		render(<Tree value={['test']} level={1} />);
+		expect(screen.getByRole('list')).toBeVisible();
+		expect(screen.getByRole('listitem')).toBeVisible();
+		expect(screen.getByRole('list')).toHaveAttribute('data-level', '1');
 	});
-	it('should return a TreeNodeList with a level > 0 and no node border', () => {
-		const wrapper = shallow(<Tree {...props} value={[]} level={1} withNodeBorder={false} />);
-		expect(wrapper.getElement()).toMatchSnapshot();
+	it('should return a list with a level > 0 and no node border', () => {
+		render(<Tree value={['test']} level={1} withNodeBorder={false} />);
+		expect(screen.getByRole('list')).toBeVisible();
+		expect(screen.getByRole('listitem')).toBeVisible();
 	});
 });