+++ packages/components/src/DataViewer/ModelViewer/Branch/ModelViewerBranch.component.js
@@ -62,6 +62,7 @@ export default class ModelViewerBranch extends Component {
 			value: this.props.value,
 		};
 		const union = this.props.isUnion(this.props.value);
+		console.log('#### union', union, this.props.isUnion());
 		if (this.state.firstClickUnion && union) {
 			this.setState(
 				{ firstClickUnion: false },
@@ -91,6 +92,7 @@ export default class ModelViewerBranch extends Component {
 			>
 				<span className={classNames(theme['tc-model-branch-content'], 'tc-model-branch-content')}>
 					<button
+						data-testid="model-branch-button"
 						className={classNames(theme['tc-model-branch-button'], 'tc-model-branch-button', {
 							[theme['tc-model-branch-button-highlighted']]: jsonpath === jsonPathSelection,
 							'tc-model-branch-button-highlighted': jsonpath === jsonPathSelection,