+++ packages/components/src/DataViewer/ModelViewer/ModelViewer.component.test.js
@@ -1,9 +1,9 @@
-import { shallow } from 'enzyme';
+import { render } from '@testing-library/react';
 import Component from './ModelViewer.component';
 
 describe('ModelViewer', () => {
 	it('should render the ModelViewer', () => {
-		const wrapper = shallow(<Component />);
-		expect(wrapper.getElement()).toMatchSnapshot();
+		const { container } = render(<Component value={[]} />);
+		expect(container.firstChild).toMatchSnapshot();
 	});
 });