+++ packages/design-system/src/components/EmptyState/index.ts
@@ -1,7 +1,7 @@
+import EmptyState from './EmptyState';
 import EmptyStateLarge from './variants/EmptyStateLarge';
 import EmptyStateMedium from './variants/EmptyStateMedium';
 import EmptyStateSmall from './variants/EmptyStateSmall';
-import EmptyState from './EmptyState';
 
 export default EmptyState;
 export { EmptyStateLarge, EmptyStateMedium, EmptyStateSmall };